improve account-workflow

+++ shepherd_server/shepherd_server/api_accounts/router.py
@@ -49,7 +49,8 @@ async def update_user(
     if isinstance(new_email, str) and new_email != user.email:
         if await User.by_email(new_email) is not None:
             raise HTTPException(406, "Email already exists")
-        user.update_email(new_email)
+        raise HTTPException(406, "Changing email currently not supported")
+        # await user.update_email(new_email)
     user = user.model_copy(update=fields)
     await user.save()
     return user
@@ -63,6 +64,10 @@ async def delete_user(
     xp_states = await WebExperiment.get_all_states(user)
     for xp_id in xp_states:
         xp = await WebExperiment.get_by_id(xp_id)
+        if xp is None:
+            raise HTTPException(
+                406, "Unexpected error while deleting experiments that do not exist"
+            )
         await ExperimentStats.update_with(xp, to_be_deleted=True)
         await xp.delete_content()
         await xp.delete()
@@ -110,9 +115,9 @@ async def forgot_password(
     user = await User.by_email(email)
     if user is None:
         return Response(status_code=200)
-    user.token_pw_reset = calculate_hash(user.email + str(local_now()))[-12:]
+    user.token_pw_reset = calculate_hash(user.email + str(local_now()))[-12:]  # => unstable token
     await get_mail_engine().send_password_reset_email(email, user.token_pw_reset)
-    await user.save_changes()
+    await user.save_changes()  # mail is sent first!
     return Response(status_code=200)
 
 
@@ -127,6 +132,7 @@ async def reset_password(
     if user is None:
         raise HTTPException(404, "Invalid password reset token")
     user.password_hash = calculate_password_hash(password)
+    user.token_pw_reset = None
     await user.save_changes()
     return user
 
@@ -147,9 +153,15 @@ async def approve(
     user = await User.by_email(email)
     if user is not None:
         raise HTTPException(409, "Account already exists")
-    token_verification = calculate_hash(email)[-12:]
-    await get_mail_engine().send_approval_email(email, token_verification)
-    return Response(status_code=200, content=token_verification)
+    user = User(
+        email=email,
+        password_hash="",  # placeholder not usable for long
+        token_verification=calculate_hash(email)[-12:],  # without date -> stable token
+        role=UserRole.user,
+    )
+    await get_mail_engine().send_approval_email(email, user.token_verification)
+    await User.insert_one(user)  # mail is sent first!
+    return Response(status_code=200, content=user.token_verification)
 
 
 @router.post("/register")
@@ -160,23 +172,19 @@ async def user_registration(
 
     To avoid spam / misuse, a valid token (generated from email) is needed.
     """
-    token_sys = calculate_hash(user_reg.email)[-12:]
-    if token_sys != user_reg.token:
-        raise HTTPException(404, "Invalid user registration token")
     user = await User.by_email(user_reg.email)
-    if user is not None:
-        raise HTTPException(409, "User with that email already exists")
-    pw_hash = calculate_password_hash(user_reg.password)
-    user = User(
-        email=user_reg.email,
-        password_hash=pw_hash,
-        disabled=False,
-        email_confirmed_at=local_now(),
-        token_verification=None,
-        role=UserRole.user,
-    )
-    await user.create()
+    if user_reg.token is None:
+        raise HTTPException(404, "Invalid registration-token")
+    if (user is None) or (user.token_verification != user_reg.token):
+        raise HTTPException(404, "Invalid account registration")
+    if user.email_confirmed_at is not None:
+        raise HTTPException(404, "Invalid user registration - account is already confirmed")
+    user.password_hash = calculate_password_hash(user_reg.password)
+    user.disabled = False
+    user.email_confirmed_at = local_now()
+    user.token_verification = None
     await get_mail_engine().send_registration_complete_email(user_reg.email)
+    await user.save_changes()  # mail is sent first!
     return user
 
 
@@ -200,7 +208,7 @@ async def verify_email(
     user.token_verification = None
     user.disabled = False
     await get_mail_engine().send_registration_complete_email(user.email)
-    await user.save_changes()
+    await user.save_changes()  # mail is sent first!
     return Response(status_code=200, content="Verification successful")
 
 
@@ -217,6 +225,6 @@ async def change_state(
     if user is None:
         raise HTTPException(status_code=401, detail="Incorrect username")
     user.disabled = not enabled
-    user.save_changes()
+    await user.save_changes()
     # TODO: inform user about it?
     return Response(status_code=200, content="State-Change successful")
